pyproject.toml
[tool.poetry]
name = "plugin-hub-ai"
version = "0.1.0"
description = ""
authors = ["Askar Yusupov <devex.soft@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.scripts]
prepare = "src.scripts:prepare"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
requests = "^2.30.0"
python-dotenv = "^1.0.0"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
lancedb = "^0.1.2"
pandas = "^2.0.1"
torch = { url = "https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.0.1%2Bcpu.cxx11.abi-cp311-cp311-linux_x86_64.whl" }
jinja2 = "^3.1.2"
sentence-transformers = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
